RAG-service/requirements.txt
fastapi
openai
google-cloud-firestore
numpy
requests
uvicorn
